Refactoring du code de MovementPhysics en plusieurs classes pour séparer les responsabilités update direction et update position en différents modules
