+++ assets/maps/2.txt
@@ -11,7 +11,7 @@ X.XXX..XX....XXX.......XX...X
 X..X..XX..XX..X..XXX.XXX..XXX
 XX...X.*.XX..XXX........X.*.X
 X..X.X.X....XX.XX.XXXXX...X.X
-X.XX.X.X.XXX.X.X..X.X.X.X.X.X
+X.XX.X.X.XXX.X.X.XX.X.X.X.X.X
 X                           X
 X.XXXXXX XXXXX XXXXX XXXXXX.X
 X......  XGGGGGGGGGX  ......X
