+++ src/GameMain.java
@@ -1,14 +1,11 @@
 package src;
 
-import src.engine.graphics.Color;
-import src.engine.graphics.Drawer;
 import src.engine.graphics.GameRenderer;
 import src.engine.input.GameInput;
 import src.engine.input.Input;
 import src.engine.physics.MovementPhysics;
-import src.engine.physics.MovingRandom;
+import src.engine.ai.MovingRandom;
 import src.entities.moving.*;
-import src.entities.space.Tile;
 import src.entities.space.TileMap;
 import src.entities.space.TileTeleport;
 import src.loaders.LevelLoader;
@@ -33,9 +30,9 @@ public class GameMain {
                 if (tileMap.get(rowIndex, columnIndex).isPacmanSpawnTile()) {
                     for (MovingEntity movingEntity : GameState.currentEntities) {
                         if (movingEntity.getEntityType() == MovingEntityType.PACMAN) {
-                            movingEntity.setTileX(columnIndex); /*Coordonnées entité par rapport à tile*/
+                            movingEntity.setTileX(columnIndex);
                             movingEntity.setTileY(rowIndex);
-                            movingEntity.setPosX( /*Coordonnées entité X en pixel ??*/
+                            movingEntity.setPosX(
                                     GameState.currentLevelPlayed.getLevelScreenOffsetLeft() +
                                     GameState.currentLevelPlayed.getTileWidth() * columnIndex +
                                     GameState.currentLevelPlayed.getTileWidth() / 2
@@ -45,7 +42,7 @@ public class GameMain {
                                     GameState.currentLevelPlayed.getTileHeight() * rowIndex +
                                     GameState.currentLevelPlayed.getTileHeight() / 2
                             );
-                            movingEntity.setTileX(columnIndex); //TODO : REDONDANT ?
+                            movingEntity.setTileX(columnIndex);
                             movingEntity.setTileY(rowIndex);
                             System.out.println(movingEntity.getPosX() + " " + movingEntity.getPosY());
                         }
@@ -69,8 +66,7 @@ public class GameMain {
                             movingEntity.setTileX(columnIndex);
                             movingEntity.setTileY(rowIndex);
                             //TODO : WARNING CODE DE TEST A PARTIR DE MAINTENANT A ENLEVER IMPERATIVEMENT PLUS TARD
-
-                            movingEntity.setCurrentDirection(Direction.LEFT); /*Direction initiale du fantôme*/
+                            movingEntity.setCurrentDirection(Direction.LEFT);
                             ArrayDeque<Input> inputInitPourTest = new ArrayDeque<>();
                             inputInitPourTest.push(Input.UP);
                             ((Ghost)movingEntity).setInputs(inputInitPourTest);
@@ -83,27 +79,19 @@ public class GameMain {
 
     public static void update(double deltaTime) {
 
-        if (newLevel) { /*TODO : A quoi sert newLevel ??*/
-            GameState.currentLevelPlayed = LevelLoader.levels.get(1); /*Ici on récup level courant ???*/
-            //DEBUG DEBUT
-            TileMap tileMap = GameState.currentLevelPlayed.getTileMap();
-            for (int i = 0 ; i < tileMap.getRowCount() ; ++i) {
-                for (int j = 0 ; j < tileMap.getColumnCount() ; ++j) {
-                    if (tileMap.get(i,j).isTeleportTile()) {
-                        TileTeleport tile = (TileTeleport)tileMap.get(i,j);
-                        System.out.println(tile.hashCode() + " " + i + " " + j + " " + tile.getTileDest().hashCode());
-                    }
-                }
-            }
-            //DEBUG FIN
+        if (newLevel) {
+            GameState.currentLevelPlayed = LevelLoader.levels.get(1);
             GameState.currentEntities.add(new Pacman(1,3.0f));
             GameState.currentEntities.add(new Pacman(2,3.0f));
             GameState.currentEntities.add(new Ghost(3.0f, null, new MovingRandom(GameState.currentLevelPlayed.getTileMap())));
+            GameState.currentEntities.add(new Ghost(5.0f, null, new MovingRandom(GameState.currentLevelPlayed.getTileMap())));
+            GameState.currentEntities.add(new Ghost(1.0f, null, new MovingRandom(GameState.currentLevelPlayed.getTileMap())));
+            GameState.currentEntities.add(new Ghost(10.0f, null, new MovingRandom(GameState.currentLevelPlayed.getTileMap())));
             initEntitiesPosition();
             newLevel = false;
         }
 
-        List<Pacman> pacmans = MovingEntity.findPacmanEntities(GameState.currentEntities); /*Direction initiale du Pacman, Immobile par défaut ??*/
+        List<Pacman> pacmans = MovingEntity.findPacmanEntities(GameState.currentEntities);
         for (Pacman pacman : pacmans) {
             if (pacman.getId() == 1) {
                 pacman.setInput(GameInput.getInputFirst());
